adicionando exceptions e bean validation nos DTO

+++ desafio-matheusLeandro/dev.matheuslf.desafio.inscritos-main/src/main/java/dev/matheuslf/desafio/inscritos/services/TaskService.java
@@ -4,6 +4,7 @@ import dev.matheuslf.desafio.inscritos.entities.DTOS.FiltersDTO;
 import dev.matheuslf.desafio.inscritos.entities.DTOS.TaskRequestDTO;
 import dev.matheuslf.desafio.inscritos.entities.DTOS.TaskResponseDTO;
 import dev.matheuslf.desafio.inscritos.entities.Task;
+import dev.matheuslf.desafio.inscritos.exceptions.NotFoundException;
 import dev.matheuslf.desafio.inscritos.repositories.ProjectRepository;
 import dev.matheuslf.desafio.inscritos.repositories.TaskRepository;
 import org.springframework.stereotype.Service;
@@ -14,11 +15,9 @@ import java.util.List;
 public class TaskService {
 
     private final TaskRepository taskRepository;
-    private final ProjectRepository projectRepository;
 
-    public TaskService(TaskRepository taskRepository, ProjectRepository projectRepository) {
+    public TaskService(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
-        this.projectRepository = projectRepository;
     }
 
     public List<TaskResponseDTO> findTasksByFilter(FiltersDTO data){
@@ -33,7 +32,7 @@ public class TaskService {
     }
 
     public TaskResponseDTO updateTask(TaskRequestDTO data, Long id){
-        Task updatedTask = taskRepository.findById(id).orElseThrow();
+        Task updatedTask = taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Not exist task with this ID"));
         Task updateTask = new Task(data);
         updatedTask.setStatus(updateTask.getStatus());
         taskRepository.save(updatedTask);
@@ -41,6 +40,7 @@ public class TaskService {
     }
 
     public void deleteTask(Long id){
+        if(!taskRepository.existsById(id)) throw new NotFoundException("Not exist task with this ID");
         taskRepository.deleteById(id);
     }
